% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metrics.R
\name{AGTP}
\alias{AGTP}
\title{Global mean temperature Change of albedo and GHG}
\usage{
AGTP(RF, duration, effect)
}
\arguments{
\item{RF}{a value of a vector of values of albedo or GHG radiaive forcings}

\item{duration}{An integer guiding the  number of years for duration of emission scenario}

\item{effect}{A string representing the effect e.g "albedo", "CO2" for carbon dioxyde etc...}
}
\value{
A vector of values corresponding to the AGTP for a time horizon defined
}
\description{
Global mean temperature Change of albedo and GHG
}
\examples{
AGTP(RF = -3, duration = 30, effect = "albedo")
AGTP(RF = -3, duration = 30, effect = "N2O")
}
