% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metrics.R
\name{GWP}
\alias{GWP}
\title{Global warming potential metric for GHG}
\usage{
GWP(x, IPCC_factor)
}
\arguments{
\item{x}{A numeric value or vector of numeric values of a GHG emission expressed into CO2 e.g 44/12 to expressed mass of C to mass of CO2 and 44/28 to convert from Mass N to mass of N2O}

\item{IPCC_factor}{A value corresponding to IPCC conversion factor of a GHG e.g 1 for CO2, 28 for CH4 and 273 for N2O at 100 years time horizon (TH) refer to AR6 (IPCC; 2023)}
}
\value{
Global warming potential values expressed in kg CO2 eq for a given (TH)
}
\description{
Global warming potential metric for GHG
}
\examples{
GWP(-280.3,273) # GWP of N2O for 100 years TH
GWP(-120.12,28) # GWP of CH4 for 100 years TH
}
