% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metrics.R
\name{GWP_albedo}
\alias{GWP_albedo}
\title{Global warming potential of albedo from Bright et al. 2016}
\usage{
GWP_albedo(RF, TH)
}
\arguments{
\item{RF}{a value of a vector of values of albedo or GHG radiaive forcings}

\item{TH}{An integer representing the time horizon, i.e. evaluation period}
}
\value{
A values in Kg CO2 eq per ha integrating the GWP of albedo of a given TH e.g. 20, 100
}
\description{
Global warming potential of albedo from Bright et al. 2016
}
\examples{
GWP_albedo(-3, 20)
GWP_labedo(-3,100)
}
